complete all required interaction logic
