+++ src/components/atoms/Svgs/index.jsx
@@ -30,7 +30,7 @@ export const ArrowDownT1Icon = (props) => (
     />
     <path
       d="m5.967 2.72 4.346 4.347a1.324 1.324 0 0 1 0 1.866L5.967 13.28"
-      stroke="#5444F2"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -87,7 +87,7 @@ export const AuditLogs = (props) => (
   >
     <path
       d="M16.1 8.333H3.908A2.239 2.239 0 0 1 1.675 6.1V3.908a2.239 2.239 0 0 1 2.233-2.233H16.1a2.239 2.239 0 0 1 2.233 2.233V6.1A2.244 2.244 0 0 1 16.1 8.333ZM16.1 18.333H3.908A2.239 2.239 0 0 1 1.675 16.1v-2.192a2.239 2.239 0 0 1 2.233-2.233H16.1a2.239 2.239 0 0 1 2.233 2.233V16.1a2.244 2.244 0 0 1-2.233 2.233ZM5 4.167v1.666M8.333 4.167v1.666M5 14.167v1.666M8.333 14.167v1.666M11.667 5H15M11.667 15H15"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -105,7 +105,7 @@ export const ComplianceIcon = (props) => (
   >
     <path
       d="M8.333 5h3.334c1.666 0 1.666-.833 1.666-1.667 0-1.666-.833-1.666-1.666-1.666H8.333c-.833 0-1.666 0-1.666 1.666C6.667 5 7.5 5 8.333 5Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -113,7 +113,7 @@ export const ComplianceIcon = (props) => (
     />
     <path
       d="M13.333 3.35c2.775.15 4.167 1.175 4.167 4.983v5c0 3.334-.833 5-5 5h-5c-4.167 0-5-1.666-5-5v-5c0-3.8 1.392-4.833 4.167-4.983"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -133,7 +133,7 @@ export const CopyIcon = (props) => (
     <path
       d="M13.333 10.75v3.5c0 2.917-1.166 4.083-4.083 4.083h-3.5c-2.917 0-4.083-1.166-4.083-4.083v-3.5c0-2.917 1.166-4.083 4.083-4.083h3.5c2.917 0 4.083 1.166 4.083 4.083Z"
       fill="#fff"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -141,7 +141,7 @@ export const CopyIcon = (props) => (
     <path
       d="M18.333 5.75v3.5c0 2.917-1.166 4.083-4.083 4.083h-.917V10.75c0-2.917-1.166-4.083-4.083-4.083H6.667V5.75c0-2.917 1.166-4.083 4.083-4.083h3.5c2.917 0 4.083 1.166 4.083 4.083Z"
       fill="#fff"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -159,7 +159,7 @@ export const CustomersIcon = (props) => (
   >
     <path
       d="M7.633 9.058a1.515 1.515 0 0 0-.275 0A3.683 3.683 0 0 1 3.8 5.367c0-2.042 1.65-3.7 3.7-3.7a3.696 3.696 0 0 1 .133 7.392ZM13.675 3.333a2.915 2.915 0 0 1 2.917 2.917 2.92 2.92 0 0 1-2.809 2.917.941.941 0 0 0-.216 0M3.467 12.133c-2.017 1.35-2.017 3.55 0 4.892 2.291 1.533 6.05 1.533 8.341 0 2.017-1.35 2.017-3.55 0-4.892-2.283-1.525-6.041-1.525-8.341 0ZM15.283 16.667a4.032 4.032 0 0 0 1.634-.725c1.3-.975 1.3-2.584 0-3.559-.459-.35-1.017-.583-1.609-.716"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -177,7 +177,7 @@ export const EyeIcon = (props) => (
   >
     <path
       d="M12.983 10A2.98 2.98 0 0 1 10 12.983 2.98 2.98 0 0 1 7.017 10 2.98 2.98 0 0 1 10 7.017 2.98 2.98 0 0 1 12.983 10Z"
-      stroke="#C1C1C1"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -202,14 +202,14 @@ export const FormatSquareIcon = (props) => (
   >
     <path
       d="M13.333 4.633v6.734a1.667 1.667 0 0 0-1.967 1.967H4.634a1.667 1.667 0 0 0-1.967-1.967V4.633a1.667 1.667 0 0 0 1.967-1.967h6.734a1.667 1.667 0 0 0 1.967 1.967Z"
-      stroke="#1A72F7"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M4.667 3a1.667 1.667 0 1 1-3.335-.001A1.667 1.667 0 0 1 4.667 3ZM14.667 3a1.667 1.667 0 1 1-3.334 0 1.667 1.667 0 0 1 3.334 0ZM4.667 13a1.667 1.667 0 1 1-3.333 0 1.667 1.667 0 0 1 3.333 0ZM14.667 13a1.667 1.667 0 1 1-3.335-.001 1.667 1.667 0 0 1 3.335.001Z"
-      stroke="#1A72F7"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -227,14 +227,14 @@ export const LiveModeIcon = (props) => (
   >
     <path
       d="M8.333 3.333h3.334A6.665 6.665 0 0 1 18.333 10a6.665 6.665 0 0 1-6.666 6.667H8.333A6.665 6.665 0 0 1 1.667 10a6.665 6.665 0 0 1 6.666-6.667Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M8.333 13.333a3.333 3.333 0 1 0 0-6.666 3.333 3.333 0 0 0 0 6.666Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -273,7 +273,7 @@ export const LogoIcon = (props) => (
       >
         <stop stopColor="#00B0FF" />
         <stop offset={0.525} stopColor="#3D5AFE" />
-        <stop offset={1} stopColor="#5444F2" />
+        <stop offset={1} stopColor={props.color || "#65717C"} />
       </linearGradient>
     </defs>
   </svg>
@@ -305,7 +305,7 @@ export const OrdersIcon = (props) => (
   >
     <path
       d="M1.667 1.667h1.45c.9 0 1.608.775 1.533 1.666l-.692 8.3a2.33 2.33 0 0 0 2.325 2.525h8.875c1.2 0 2.25-.983 2.342-2.175l.45-6.25c.1-1.383-.95-2.508-2.342-2.508H4.85M13.542 18.333a1.042 1.042 0 1 0 0-2.083 1.042 1.042 0 0 0 0 2.083ZM6.875 18.333a1.042 1.042 0 1 0 0-2.083 1.042 1.042 0 0 0 0 2.083ZM7.5 6.667h10"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -324,7 +324,7 @@ export const OverviewIcon = (props) => (
   >
     <path
       d="M14.167 8.333h1.666c1.667 0 2.5-.833 2.5-2.5V4.167c0-1.667-.833-2.5-2.5-2.5h-1.666c-1.667 0-2.5.833-2.5 2.5v1.666c0 1.667.833 2.5 2.5 2.5ZM4.167 18.333h1.666c1.667 0 2.5-.833 2.5-2.5v-1.666c0-1.667-.833-2.5-2.5-2.5H4.167c-1.667 0-2.5.833-2.5 2.5v1.666c0 1.667.833 2.5 2.5 2.5ZM5 8.333a3.333 3.333 0 1 0 0-6.666 3.333 3.333 0 0 0 0 6.666ZM15 18.333a3.333 3.333 0 1 0 0-6.666 3.333 3.333 0 0 0 0 6.666Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -343,14 +343,14 @@ export const ProductsIcon = (props) => (
   >
     <path
       d="M7 5.417h6c2.833 0 3.117 1.325 3.308 2.941l.75 6.25c.242 2.05-.391 3.725-3.308 3.725H6.258c-2.925 0-3.558-1.675-3.308-3.725l.75-6.25c.183-1.616.467-2.941 3.3-2.941Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M6.667 6.667V3.75c0-1.25.833-2.083 2.083-2.083h2.5c1.25 0 2.083.833 2.083 2.083v2.917M17.008 14.192H6.667"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -369,7 +369,7 @@ export const RowVerticalIcon = (props) => (
     <path
       d="M13.267 9H2.733c-1 0-1.4.427-1.4 1.487v2.693c0 1.06.4 1.487 1.4 1.487h10.534c1 0 1.4-.427 1.4-1.487v-2.693c0-1.06-.4-1.487-1.4-1.487ZM13.267 1.333H2.733c-1 0-1.4.427-1.4 1.487v2.693c0 1.06.4 1.487 1.4 1.487h10.534c1 0 1.4-.427 1.4-1.487V2.82c0-1.06-.4-1.487-1.4-1.487Z"
       fill="#fff"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -387,7 +387,7 @@ export const SearchNormalIcon = (props) => (
   >
     <path
       d="M7.667 14a6.333 6.333 0 1 0 0-12.667 6.333 6.333 0 0 0 0 12.667ZM14.667 14.667l-1.334-1.334"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -405,7 +405,7 @@ export const SettingsT1Icon = (props) => (
   >
     <path
       d="M10 12.5a2.5 2.5 0 1 0 0-5 2.5 2.5 0 0 0 0 5Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -413,7 +413,7 @@ export const SettingsT1Icon = (props) => (
     />
     <path
       d="M1.667 10.733V9.267c0-.867.708-1.584 1.583-1.584 1.508 0 2.125-1.066 1.367-2.375A1.583 1.583 0 0 1 5.2 3.15l1.442-.825c.658-.392 1.508-.158 1.9.5l.091.158c.75 1.309 1.984 1.309 2.742 0l.092-.158c.391-.658 1.241-.892 1.9-.5l1.441.825a1.583 1.583 0 0 1 .584 2.158c-.759 1.309-.142 2.375 1.366 2.375.867 0 1.584.709 1.584 1.584v1.466c0 .867-.709 1.584-1.584 1.584-1.508 0-2.125 1.066-1.366 2.375a1.58 1.58 0 0 1-.584 2.158l-1.441.825c-.659.392-1.509.158-1.9-.5l-.092-.158c-.75-1.309-1.983-1.309-2.742 0l-.091.158c-.392.658-1.242.892-1.9.5L5.2 16.85a1.582 1.582 0 0 1-.583-2.158c.758-1.309.141-2.375-1.367-2.375a1.588 1.588 0 0 1-1.583-1.584Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -432,7 +432,7 @@ export const SettingsT2Icon = (props) => (
   >
     <path
       d="M6.832 15.75h4.328c1.59 0 1.59 0 2.602-1.5l2.386-4.125c.36-.623.36-1.635 0-2.25L13.762 3.75c-1.012-1.5-1.012-1.5-2.595-1.5H6.833c-1.582 0-1.582 0-2.595 1.5L1.852 7.875c-.36.615-.36 1.627 0 2.25l2.385 4.125c1.013 1.5 1.013 1.5 2.595 1.5Z"
-      stroke="#5444F2"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -440,7 +440,7 @@ export const SettingsT2Icon = (props) => (
     <path
       d="M11.25 9a2.25 2.25 0 1 0-4.5 0 2.25 2.25 0 0 0 4.5 0Z"
       fill="#fff"
-      stroke="#5444F2"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -458,7 +458,7 @@ export const TransactionsIcon = (props) => (
   >
     <path
       d="M2.983 4.3h11.534c1.383 0 2.5 1.117 2.5 2.5v2.767"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -466,7 +466,7 @@ export const TransactionsIcon = (props) => (
     />
     <path
       d="M5.617 1.667 2.983 4.3l2.634 2.633M17.017 15.7H5.483a2.497 2.497 0 0 1-2.5-2.5v-2.767"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -474,7 +474,7 @@ export const TransactionsIcon = (props) => (
     />
     <path
       d="m14.383 18.333 2.634-2.633-2.634-2.633"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -493,13 +493,13 @@ export const VirtualAccountsIcon = (props) => (
   >
     <path
       d="m18.058 11.917-.333 4.166c-.125 1.275-.225 2.25-2.483 2.25H4.758c-2.258 0-2.358-.975-2.483-2.25l-.333-4.166a2.526 2.526 0 0 1 .541-1.825l.017-.017a2.479 2.479 0 0 1 1.925-.908h11.15c.775 0 1.458.35 1.908.891.009.009.017.017.017.025.408.492.633 1.134.558 1.834Z"
-      stroke="#5444F2"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
     />
     <path
       d="M2.917 9.525V5.233c0-2.833.708-3.541 3.541-3.541h1.059c1.058 0 1.3.316 1.7.85l1.058 1.416c.267.35.425.567 1.133.567h2.125c2.834 0 3.542.708 3.542 3.542v1.491M7.858 14.167h4.284"
-      stroke="#5444F2"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -518,28 +518,28 @@ export const WalletIcon = (props) => (
   >
     <path
       d="m10.55 2.098-.025.059-2.417 5.608H5.733c-.566 0-1.108.117-1.6.325l1.459-3.483.033-.084.058-.133c.017-.05.034-.1.059-.142 1.091-2.525 2.325-3.1 4.808-2.15Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M15.042 7.932a3.967 3.967 0 0 0-1.175-.167H8.108l2.417-5.608.025-.059c.125.042.242.1.367.15l1.841.775c1.025.425 1.742.867 2.175 1.4.084.1.15.192.209.3.075.117.133.234.166.359.034.075.059.15.075.216.225.7.092 1.559-.341 2.634Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M17.935 11.832v1.625c0 .167-.008.333-.017.5-.158 2.908-1.783 4.375-4.866 4.375h-6.5c-.2 0-.4-.017-.592-.042-2.65-.175-4.067-1.591-4.242-4.241a4.597 4.597 0 0 1-.042-.592v-1.625A4.078 4.078 0 0 1 4.143 8.09a4.127 4.127 0 0 1 1.6-.325h8.133c.409 0 .809.059 1.176.167a4.086 4.086 0 0 1 2.883 3.9Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
     />
     <path
       d="M5.592 4.607 4.133 8.09a4.078 4.078 0 0 0-2.466 3.742V9.39a4.874 4.874 0 0 1 3.925-4.783ZM17.932 9.39v2.441A4.066 4.066 0 0 0 15.05 7.94c.433-1.084.558-1.934.35-2.642a.946.946 0 0 0-.075-.217 4.853 4.853 0 0 1 2.608 4.309Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeLinecap="round"
       strokeLinejoin="round"
@@ -557,7 +557,7 @@ export const WelcomeIcon = (props) => (
   >
     <path
       d="M5 3.333A8.336 8.336 0 0 1 18.333 10c0 4.6-3.733 8.333-8.333 8.333A8.336 8.336 0 0 1 1.667 10c0-1.508.4-2.925 1.108-4.15L10 10"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -565,7 +565,7 @@ export const WelcomeIcon = (props) => (
     />
     <path
       d="M5.692 7.467A4.934 4.934 0 0 0 5 10c0 2.758 2.242 5 5 5s5-2.242 5-5-2.242-5-5-5c-.758 0-1.483.167-2.125.475"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
@@ -584,7 +584,7 @@ export const NotificationIcon = (props) => (
   >
     <path
       d="M8.018 1.425c-2.759 0-5 2.242-5 5v2.408c0 .509-.217 1.284-.475 1.717l-.959 1.592c-.591.983-.183 2.075.9 2.441 3.592 1.2 7.467 1.2 11.059 0a1.669 1.669 0 0 0 .9-2.441l-.959-1.592c-.25-.433-.466-1.208-.466-1.717V6.425c0-2.75-2.25-5-5-5Z"
-      stroke="#65717C"
+      stroke={props.color || "#65717C"}
       strokeWidth={1.5}
       strokeMiterlimit={10}
       strokeLinecap="round"
