+++ src/components/molecules/SideNav/index.jsx
@@ -33,6 +33,10 @@ const LinkItem = styled(Link)`
   display: block;
   display: flex;
   align-items: center;
+
+  &:hover {
+    font-weight: 600;
+  }
 `;
 const TopLinks = styled.div`
   & div:not(:last-child) {
@@ -41,13 +45,14 @@ const TopLinks = styled.div`
 
   @media only screen and (${screen.sm}) {
     display: grid;
-    gap: 20px;
+    gap: 10px;
     grid-template-columns: auto auto auto;
     padding: 10px;
     width: 100%;
   }
 `;
 const BottomLinks = styled.div`
+  background-color: "#F5F6FA33";
   & div:not(:last-child) {
     padding: 16px 0;
   }
@@ -65,6 +70,10 @@ const LinkHeading = styled(Text)`
   font-weight: 600;
   padding: 6px 0 16px 0;
   margin: 0;
+
+  @media only screen and (${screen.sm}) {
+    display: none;
+  }
 `;
 const LinkTextWrapper = styled.span`
   display: block;
@@ -135,7 +144,17 @@ export const SideNav = () => {
           <Flex>
             <LinkItem>
               <VirtualAccountsIcon />
-              <LinkTextWrapper>Virtual Account</LinkTextWrapper>
+              <Flex direction="column">
+                <LinkTextWrapper>Virtual Account</LinkTextWrapper>
+                <Text
+                  fontWeight="500"
+                  color={baniColors.green}
+                  fontSize={typeScale.smaller}
+                  margin="5px 0 0 10px"
+                >
+                  New
+                </Text>
+              </Flex>
             </LinkItem>
           </Flex>
         </div>
