+++ src/components/organisms/index.js
@@ -1,1 +1,3 @@
 export * from "./Table";
+export * from "./DashboardActivityFeed";
+export * from "./VirtualAccounts";
